refactor starter module.
