+++ shenyu-spring-boot-starter/pom.xml
@@ -33,6 +33,7 @@
         <module>shenyu-spring-boot-starter-client</module>
         <module>shenyu-spring-boot-starter-registry</module>
         <module>shenyu-spring-boot-starter-sdk</module>
+        <module>shenyu-spring-boot-starter-sdk-feign</module>
         <module>shenyu-spring-boot-starter-k8s</module>
     </modules>
