+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-sdk-feign/src/main/resources/META-INF/spring.factories
