+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-sdk/src/main/resources/META-INF/spring.factories
@@ -15,4 +15,5 @@
 # limitations under the License.
 #
 
-provides: shenyu-spring-boot-starter-sdk-spring
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.apache.shenyu.springboot.starter.sdk.ShenyuSdkAutoConfiguration
