+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-sdk/src/main/resources/META-INF/spring.provides
@@ -15,5 +15,4 @@
 # limitations under the License.
 #
 
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.apache.shenyu.springboot.starter.sdk.spring.ShenyuSdkAutoConfiguration
+provides: shenyu-spring-boot-starter-sdk
